+++ cpp/unittest/gis/cuda/functor_point_tests.cpp
@@ -22,7 +22,7 @@
 #include "gis/cuda/common/gis_definitions.h"
 #include "gis/cuda/functor/st_distance.h"
 #include "gis/cuda/functor/st_point.h"
-#include "gis/test_common/test_common.h"
+#include "gis/cuda/test_common/test_common.h"
 using std::vector;
 namespace arctern {
 namespace gis {
