+++ cpp/unittest/gis/cuda/functor_within_tests.cpp
@@ -21,7 +21,7 @@
 
 #include "gis/cuda/common/gis_definitions.h"
 #include "gis/cuda/functor/st_within.h"
-#include "gis/test_common/test_common.h"
+#include "gis/cuda/test_common/test_common.h"
 
 namespace arctern {
 namespace gis {
